fix(io): harden parsers and sniffers against real-world files
An import-layer edge-case audit (malformed/truncated files, sniffer
mis-routing) found six issues; all fixes are parity-safe on the existing
corpus (golden tests unchanged):

- xrdml: sort scans by float(appendNumber) — int("1.0") crashed on the
  float-formatted appendNumber some exporters write.
- qd.import_ppms: the PPMS sniffer accepts any QD-ish plain CSV (e.g.
  resistance-vs-temperature, moment-only), so the default x/y "field"/"moment"
  may be absent — degrade to auto-detected columns instead of KeyError.
- refl1d sniffer: match column signals on word boundaries (\brho\b, \bq\b +
  \br\b) so prose like "rhodium thermometer" / "quick readout" no longer
  false-positives and mis-routes PPMS files into the refl1d parser.
- refl1d parser: pad/truncate ragged rows to NaN (MATLAB textscan parity)
  instead of crashing np.asarray on inhomogeneous shape; raise a clear error
  on a single-column file (was a silent 0-channel DataStruct).
- ncnr: scan the whole comment header for intensity/background metadata, not
  just the first 5 lines.

+11 regression tests; committed fixture refl1d_refl_fit.dat already present.

+++ src/quantized/io/xrdml.py
@@ -116,7 +116,9 @@ def import_xrdml(filepath: str | Path, *, intensity: str = "cps") -> DataStruct:
     scans_xml = _find_all(root, "scan")
     if not scans_xml:
         raise ValueError(f"no <scan> element in {path.name}")
-    scans_xml.sort(key=lambda s: int(s.get("appendNumber", "1")))
+    # appendNumber is usually an int but some exporters write "1.0"; float()
+    # accepts both and preserves ordering (int() raised on float strings).
+    scans_xml.sort(key=lambda s: float(s.get("appendNumber", "1")))
 
     wavelength = _wavelength(root)
     counting_time = float("nan")
